refactor: move error handling from extension to AppError class
