+++ ARCHITECTURE.md
@@ -14,7 +14,6 @@
  3. Action層: 状態変更の意図を表現（Reduxパターン）
  4. Store層: 状態管理、ビジネスロジック
  5. Screens層: UI表示、ユーザーインタラクション
- 6. Extension層: 既存クラスの機能拡張
 
 ## 状態管理
 ### Reduxパターンの採用
